test(cli): isolate pr-bot archive Git state
